requirements.txt
cryptography==39.0.1
pycryptodome==3.15.0
pytest==8.3.4